refactor: チャンネル一覧のデータと取得メソッドを presenters に移動
